requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.4.2
Automat==20.2.0
bcrypt==3.2.2
billiard==3.6.4.0
celery==5.2.6
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==37.0.2
daphne==3.0.2
Deprecated==1.2.13
Django==4.0.4
django-cors-headers==3.11.0
django-debug-toolbar==3.4.0
django-filter==21.1
django-json-widget==1.1.1
django-redis==5.2.0
djangorestframework==3.13.1
drf-spectacular==0.22.1
frozenlist==1.3.0
future==0.18.2
geoip2==4.5.0
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
jsonschema==4.5.1
kombu==5.2.4
loguru==0.6.0
maxminddb==2.2.0
multidict==6.0.2
mysqlclient==2.1.0
packaging==21.3
paramiko==2.11.0
prompt-toolkit==3.0.29
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-decouple==3.6
pytz==2022.1
PyYAML==6.0
redis==4.2.0
requests==2.27.1
sentry-sdk==1.5.12
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
yarl==1.7.2
zope.event==4.5.0
zope.interface==5.4.0
rich~=12.4.1
django-extensions
pymysql
